:recycle: Rename to router package
