+++ lib/initial_setting/initial_setting_2.dart
@@ -1,4 +1,4 @@
-import 'package:Pilll/application/router.dart';
+import 'package:Pilll/router/router.dart';
 import 'package:Pilll/store/initial_setting.dart';
 import 'package:Pilll/components/atoms/buttons.dart';
 import 'package:Pilll/components/atoms/color.dart';
