+++ lib/initial_setting/initial_setting_3.dart
@@ -1,4 +1,4 @@
-import 'package:Pilll/application/router.dart';
+import 'package:Pilll/router/router.dart';
 import 'package:Pilll/components/organisms/setting/setting_menstruation_page.dart';
 import 'package:Pilll/initial_setting/initial_setting_4.dart';
 import 'package:Pilll/store/initial_setting.dart';
